refactoring - move all relevant logic to sigle command runner
